compile site view earlier
