+++ crates/api/api/src/lib.rs
@@ -13,9 +13,10 @@ pub mod comment;
 pub mod community;
 pub mod local_user;
 pub mod post;
-pub mod reports;pub mod stuff;
+pub mod reports;
 pub mod site;
 pub mod sitemap;
+pub mod stuff;
 
 /// Converts the captcha to a base64 encoded wav audio file
 pub(crate) fn captcha_as_wav_base64(captcha: &Captcha) -> LemmyResult<String> {
