+++ crates/api/api/src/stuff/read_person.rs
@@ -1,3 +1,4 @@
+use crate::stuff::resolve_person_id_from_id_or_username;
 use activitypub_federation::config::Data;
 use actix_web::web::{Json, Query};
 use lemmy_api_utils::{
@@ -12,8 +13,6 @@ use lemmy_db_views_person::{
 };
 use lemmy_db_views_site::SiteView;
 use lemmy_utils::error::LemmyResult;
-
-use crate::stuff::resolve_person_id_from_id_or_username;
 
 pub async fn read_person(
   data: Query<GetPersonDetails>,
