+++ crates/api/api_common/src/account.rs
@@ -1,3 +1,9 @@
+pub use lemmy_db_views_person_content_combined::api::{
+  ListPersonHidden,
+  ListPersonHiddenResponse,
+  ListPersonRead,
+  ListPersonReadResponse,
+};
 pub use lemmy_db_views_person_liked_combined::{
   ListPersonLiked,
   ListPersonLikedResponse,
@@ -8,16 +14,7 @@ pub use lemmy_db_views_person_saved_combined::{
   ListPersonSavedResponse,
   PersonSavedCombinedView,
 };
-pub use lemmy_db_views_site::api::{
-  DeleteAccount,
-  ListPersonHidden,
-  ListPersonHiddenResponse,
-  ListPersonRead,
-  ListPersonReadResponse,
-  MyUserInfo,
-  SaveUserSettings,
-};
-
+pub use lemmy_db_views_site::api::{DeleteAccount, MyUserInfo, SaveUserSettings};
 pub mod auth {
   pub use lemmy_db_schema::source::login_token::LoginToken;
   pub use lemmy_db_views_registration_applications::api::Register;
