+++ crates/db_views/person_content_combined/src/lib.rs
@@ -34,6 +34,7 @@ use {
   lemmy_db_views_local_user::LocalUserView,
 };
 
+pub mod api;
 #[cfg(feature = "full")]
 pub mod impls;
