+++ src/api_routes.rs
@@ -97,6 +97,16 @@ use lemmy_api::{
       unread_count::get_unread_registration_application_count,
     },
   },
+  stuff::{
+    list_comments::{list_comments, list_comments_slim},
+    list_person_content::list_person_content,
+    list_posts::list_posts,
+    read_community::get_community,
+    read_person::read_person,
+    resolve_object::resolve_object,
+    search::search,
+    user_settings_backup::{export_settings, import_settings},
+  },
 };
 use lemmy_api_crud::{
   comment::{
@@ -156,16 +166,6 @@ use lemmy_api_crud::{
     delete::delete_account,
     my_user::get_my_user,
   },
-};
-use lemmy_api::stuff::{
-  list_comments::{list_comments, list_comments_slim},
-  list_person_content::list_person_content,
-  list_posts::list_posts,
-  read_community::get_community,
-  read_person::read_person,
-  resolve_object::resolve_object,
-  search::search,
-  user_settings_backup::{export_settings, import_settings},
 };
 use lemmy_routes::images::{
   delete::{
